// frontend/src/api/models/EnrollmentResponse.ts
/* tslint:disable */
/* eslint-disable */
/**
 * FastAPI
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * 
 * @export
 * @interface EnrollmentResponse
 */
export interface EnrollmentResponse {
    /**
     * 
     * @type {number}
     * @memberof EnrollmentResponse
     */
    studentId: number;
    /**
     * 
     * @type {number}
     * @memberof EnrollmentResponse
     */
    courseId: number;
    /**
     * 
     * @type {number}
     * @memberof EnrollmentResponse
     */
    assignerId: number;
    /**
     * 
     * @type {Date}
     * @memberof EnrollmentResponse
     */
    completedAt?: Date | null;
    /**
     * 
     * @type {Date}
     * @memberof EnrollmentResponse
     */
    enrolledAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof EnrollmentResponse
     */
    deadline?: Date | null;
    /**
     * 
     * @type {boolean}
     * @memberof EnrollmentResponse
     */
    isActive?: boolean;
    /**
     * 
     * @type {number}
     * @memberof EnrollmentResponse
     */
    enrollmentId: number;
}

/**
 * Check if a given object implements the EnrollmentResponse interface.
 */
export function instanceOfEnrollmentResponse(value: object): value is EnrollmentResponse {
    if (!('studentId' in value) || value['studentId'] === undefined) return false;
    if (!('courseId' in value) || value['courseId'] === undefined) return false;
    if (!('assignerId' in value) || value['assignerId'] === undefined) return false;
    if (!('enrollmentId' in value) || value['enrollmentId'] === undefined) return false;
    return true;
}

export function EnrollmentResponseFromJSON(json: any): EnrollmentResponse {
    return EnrollmentResponseFromJSONTyped(json, false);
}

export function EnrollmentResponseFromJSONTyped(json: any, ignoreDiscriminator: boolean): EnrollmentResponse {
    if (json == null) {
        return json;
    }
    return {
        
        'studentId': json['student_id'],
        'courseId': json['course_id'],
        'assignerId': json['assigner_id'],
        'completedAt': json['completed_at'] == null ? undefined : (new Date(json['completed_at'])),
        'enrolledAt': json['enrolled_at'] == null ? undefined : (new Date(json['enrolled_at'])),
        'deadline': json['deadline'] == null ? undefined : (new Date(json['deadline'])),
        'isActive': json['is_active'] == null ? undefined : json['is_active'],
        'enrollmentId': json['enrollment_id'],
    };
}

export function EnrollmentResponseToJSON(json: any): EnrollmentResponse {
    return EnrollmentResponseToJSONTyped(json, false);
}

export function EnrollmentResponseToJSONTyped(value?: EnrollmentResponse | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'student_id': value['studentId'],
        'course_id': value['courseId'],
        'assigner_id': value['assignerId'],
        'completed_at': value['completedAt'] == null ? undefined : ((value['completedAt'] as any).toISOString()),
        'enrolled_at': value['enrolledAt'] == null ? undefined : ((value['enrolledAt']).toISOString().substring(0,10)),
        'deadline': value['deadline'] == null ? undefined : ((value['deadline'] as any).toISOString().substring(0,10)),
        'is_active': value['isActive'],
        'enrollment_id': value['enrollmentId'],
    };
}